+++ node/resolvers/index.ts
@@ -1,0 +1,11 @@
+import {
+  fieldResolvers as affiliateOrdersFieldResolvers,
+  queries as affiliateOrdersQueries,
+} from './affiliateOrders'
+
+export const resolvers = {
+  ...affiliateOrdersFieldResolvers,
+  Query: {
+    ...affiliateOrdersQueries,
+  },
+}