+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- GraphQL Schema, typings and resolvers for Affiliate Orders
+
 ## [0.11.0] - 2022-01-20
 
 ### Added